+++ crates/database/src/profile_feeds.rs
@@ -1,5 +1,5 @@
 #[derive(Debug)]
 pub struct InsertData<'a> {
     pub profile_id: &'a str,
-    pub feed_id: i32,
+    pub feed_id: i64,
 }
